feat(anthropic-llm)!: extract AnthropicProvider into @mcp-abap-adt/anthropic-llm
- Move anthropic-provider.ts to packages/anthropic-llm/src/
- Move anthropic-provider.test.ts to packages/anthropic-llm/src/__tests__/
- Create package skeleton (package.json, tsconfig.json, README.md, index.ts)
- Update server imports to use @mcp-abap-adt/anthropic-llm
- Add anthropic-llm to server tsconfig.json references
- Back-compat re-export in server's llm-providers/index.ts (stays until Task 14)
- All tests pass: anthropic-llm (19 tests), llm-agent-server (full suite)
