+++ packages/anthropic-llm/src/index.ts
@@ -1,0 +1,2 @@
+export type { AnthropicConfig } from './anthropic-provider.js';
+export { AnthropicProvider } from './anthropic-provider.js';
