+++ packages/anthropic-llm/tsconfig.json
@@ -8,9 +8,5 @@
   },
   "include": ["src/**/*"],
   "exclude": ["**/__tests__/**", "**/*.test.ts", "dist"],
-  "references": [
-    { "path": "../llm-agent" },
-    { "path": "../openai-llm" },
-    { "path": "../anthropic-llm" }
-  ]
+  "references": [{ "path": "../llm-agent" }]
 }
